refactor: utils package
