+++ now/utils/errors/helpers.py
@@ -1,0 +1,6 @@
+class RetryException(Exception):
+    pass
+
+
+class DemoAvailableException(Exception):
+    pass
