+++ now/utils/flow/helpers.py
@@ -1,0 +1,78 @@
+import cowsay
+import yaml
+from jina.jaml import JAML
+
+from now.deployment.deployment import list_all_wolf, status_wolf
+from now.thirdparty.PyInquirer.prompt import prompt
+
+
+def get_flow_status(action, **kwargs):
+    choices = []
+    # Add all remote Flows that exists with the namespace `nowapi`
+    alive_flows = list_all_wolf(status='Serving')
+    for flow_details in alive_flows:
+        choices.append(flow_details['name'])
+    if len(choices) == 0:
+        cowsay.cow(f'nothing to {action}')
+        return
+    else:
+        questions = [
+            {
+                'type': 'list',
+                'name': 'cluster',
+                'message': f'Which cluster do you want to {action}?',
+                'choices': choices,
+            }
+        ]
+        cluster = maybe_prompt_user(questions, 'cluster', **kwargs)
+
+    flow = [x for x in alive_flows if x['name'] == cluster][0]
+    flow_id = flow['id']
+    _result = status_wolf(flow_id)
+    if _result is None:
+        print(f'❎ Flow not found in JCloud. Likely, it has been deleted already')
+    return _result, flow_id, cluster
+
+
+def get_flow_id(host):
+    return host[len('https://') : -len('-http.wolf.jina.ai')]
+
+
+def write_env_file(env_file, config):
+    config_string = '\n'.join([f'{key}={value}' for key, value in config.items()])
+    with open(env_file, 'w+') as fp:
+        fp.write(config_string)
+
+
+class Dumper(yaml.Dumper):
+    def increase_indent(self, flow=False, *args, **kwargs):
+        return super().increase_indent(flow=flow, indentless=False)
+
+
+def write_flow_file(flow_yaml_content, new_yaml_file_path):
+    with open(new_yaml_file_path, 'w') as f:
+        JAML.dump(
+            flow_yaml_content,
+            f,
+            indent=2,
+            allow_unicode=True,
+            Dumper=Dumper,
+        )
+
+
+def maybe_prompt_user(questions, attribute, **kwargs):
+    """
+    Checks the `kwargs` for the `attribute` name. If present, the value is returned directly.
+    If not, the user is prompted via the cmd-line using the `questions` argument.
+
+    :param questions: A dictionary that is passed to `PyInquirer.prompt`
+        See docs: https://github.com/CITGuru/PyInquirer#documentation
+    :param attribute: Name of the value to get. Make sure this matches the name in `kwargs`
+
+    :return: A single value of either from `kwargs` or the user cli input.
+    """
+    if kwargs and attribute in kwargs:
+        return kwargs[attribute]
+    else:
+        answer = prompt(questions)
+        return answer[attribute]
